fix tracing
In #187 we collapsed the spans which caused all the LazyPrintf in defer
statements to be dropped.

In this PR I revert #187, deduplicate trace logs (EG we logged "num
files" and `opts` in multiple places) and log `stats` in streaming
search.

Test plan:

I confirmed locally that the logs show up as expected.