boost graphql types in results
Because "type" in scip-ctags matches so much we can't boost it as high
as we do for classes/structs in our language specific boosting. In the
case of GraphQL this is useful to boost.

Test Plan: A search like "type User" in the sourcegraph codebase now
includes results from the schema.graphql files.