gomod: update go-ctags and ctags in docker
We bump go-ctags to include markdown files in symbol analysis. We
already have ranking behaviour for markdown files but due to go-ctags it
won't be taken into account (until this commit).

Additionally we bump the version included in the docker container to
v6.0.0.

Test Plan: CI and "docker build -t zoekt ."