score: experimental extension novelty in sorting (#665)
Right now we boost a file extension that hasn't been seen to the 3rd
position. This is gated by an environment variable which defaults to
on. I want to explore if there are ways we can turn on this behaviour 
with the query language.

Test Plan: go run ./cmd/zoekt foo