archive: e2e test for ranking against sourcegraph repo
This is an initial framework for having golden file results for search
results against a real repository. At first we have only added one query
and one repository, but it should be straightforward to grow this list
further.

The golden files we write to disk are a summary of results with debug
information. This matches how we have been using the zoekt CLI tool on
the keyword branch during our ranking work.

Test Plan: go test