build: ignore out of bound lines from ctags (#694)
universal-ctags sometimes returns lines that are out of bounds. In
practice it seems to only do an off by one. We haven't noticed the
linenum error until a recent change of mine which didn't append an extra
entry to NLS if the file was terminated by "\n". In practice this would
end up being filtered out later on. So we update to just continue rather
than error here.

An example is
https://github.com/sourcegraph/sourcegraph/blob/v5.2.2/client/web-sveltekit/.storybook/main.ts

  $ universal-ctags '--fields=*' --output-format=json main.ts | grep 22
  {"_type": "tag", "name": "config", "path": "main.ts", "pattern": "/^export default config$/", "language": "TypeScript", "line": 22, "kind": "constant", "roles": "def"}

  $ wc -l main.ts
  21 main.ts

  $ tail -n1 main.ts
  export default config

Test Plan: added a unit test