archive: re-use shard cache in e2e rank tests
Indexing the sourcegraph repo takes a couple of seconds on my laptop.
Most changes we make do not change the structure of the shard. This
commit introduces a default behaviour of reusing shards between go test
invocations. If on CI it will always recompute. This way if we
accidently think everything is alright locally it will get caught on CI.

Test Plan: CI