e2e: introduce optional shard cache
While iterating on the ranking tests it is frustrating to wait for the
index to be built. This commit adds a "-shard_cache" flag to reuse the
computed shards. It defaults to off since this could lead to unexpected
results if we change how we index.

Test Plan: run go test several times with the flag, only the first run
should be slow. Then confirm without flag it is "slow".

  $ go test -shard_cache
  PASS
  ok      github.com/sourcegraph/zoekt/internal/e2e       23.307s
  $ go test -shard_cache
  PASS
  ok      github.com/sourcegraph/zoekt/internal/e2e       0.235s
  $ go test -shard_cache
  PASS
  ok      github.com/sourcegraph/zoekt/internal/e2e       0.218s
  $ go test
  PASS
  ok      github.com/sourcegraph/zoekt/internal/e2e       23.416s