matchtree: factor out regexpMatchTree construction
We want to introduce more state in construction, so this ensures we
always initialize it. Additionally unify construction by always taking
in a query.Regexp. For the short strings, we rely on directly creating
the literal regex.

Note: I explored a little bit avoiding the prefix + .String stuff and
instead maybe directly taking a mutated Regexp with flags set. However,
this was a deeper change to follow up on.

Test Plan: go test