BM25: Boost file name matches at root
With this change we prioritize file name matches at the root of the
repository. This is based on the intuition that more important files
tend to be closer to the root.

We also change the parameter b in the BM25 scoring function from 0.75 to
0.3 to reduce the impact of the document length on the final score. This
is based on experiments that showed that our current scoring overly
penalizes long but important documents.

For example, we consider documents such as a README.md or CHANGELOG at
the root of the repository of high quality. However, these documents
also tend to be relatively long and are thus penalized.

Test plan:
Updated unit test