chore: remove unused feature "document ranks"
Document ranks was an experimental feature that allowed Zoekt to query
Sourcegraph for additional ranking signals for files at index-time. We
have decided to remove this feature because it was never used in
production.

Note:
Most of the changes are Sourcegraph specific. However there are some
exported fields in `SearchOptions` and `build.Options` as well as flags
for zoekt-git-index which I simply removed.

Test plan:
updated tests