tenant: add tenant to trace
This adds the tenant ID to the trace. I also move the pprof logging from
FromContext to a higher level searcher, because the number of events was
too high, because we logged missing tenants per document.

Note: Before, pprof logging didn't work at all, because we read the
tenant enforcemnt ENV after we set the pprof profile, so the profile
was always nil.

Test plan:
Checked locally that tenants show up in the traces and
zoekt_missing_tenant shows up as pprof profile.